requirements.txt
https://github.com/masoniteframework/masonite/archive/develop.zip
masonite-logging>=1.0,<2.0
masonite-validation>=3.0,<4.0
